AuftragsCache refactored, migration stuff... (all tests OK)

+++ projectforge-rest/src/main/kotlin/org/projectforge/rest/AddressPagesRest.kt
@@ -362,7 +362,7 @@ class AddressPagesRest
                       .add(
                         createFavoriteRow(
                           "isFavoriteCard",
-                          UISelect<Int>(
+                          UISelect<Long>(
                             "addressbookList", lc,
                             multi = true,
                             autoCompletion = AutoCompletion<Int>(
