+++ projectforge-rest/src/main/kotlin/org/projectforge/rest/AddressServicesRest.kt
@@ -160,7 +160,7 @@ class AddressServicesRest {
   }
 
   @GetMapping("exportVCard/{id}")
-  fun exportVCard(@PathVariable("id") id: Int?): ResponseEntity<*> {
+  fun exportVCard(@PathVariable("id") id: Long?): ResponseEntity<*> {
     val address = addressDao.find(id) ?: return ResponseEntity<Any>(HttpStatus.NOT_FOUND)
     val filename = ("ProjectForge-" + ReplaceUtils.encodeFilename(address.fullName, true) + "_"
         + DateHelper.getDateAsFilenameSuffix(Date()) + ".vcf")
