+++ projectforge-rest/src/main/kotlin/org/projectforge/rest/calendar/CalEventPagesRest.kt
@@ -150,7 +150,7 @@ class CalEventPagesRest() : AbstractDTOPagesRest<CalEventDO, CalEvent, CalEventD
         )
       }
     } else {
-      val calendarId = NumberHelper.parseInteger(request.getParameter("calendar"))
+      val calendarId = NumberHelper.parseLong(request.getParameter("calendar"))
       if (calendarId != null && calendarId > 0) {
         dto.calendar = teamCalDao.find(calendarId)
       }
